+++ src/main/java/br/edu/fatecgru/mercado_inteligente/repository/ItemCarrinhoRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import br.edu.fatecgru.mercado_inteligente.model.entity.ItemCarrinho;
 
 public interface ItemCarrinhoRepository extends JpaRepository<ItemCarrinho, Long> {
-    Optional<ItemCarrinho> findByCarrinhoIdAndProdutoId(Integer carrinhoId, Long produtoId);
+    Optional<ItemCarrinho> findByCarrinhoIdAndProdutoId(Long carrinhoId, Long produtoId);
 }
